mcp serve
