--- internal/cli/mcp_install_test.go
@@ -1,195 +1,0 @@
-package cli
-
-import (
-	"bytes"
-
-	// "context" // Removed unused import
-	"encoding/json"
-	"fmt"
-	"os"
-	"path/filepath"
-	"runtime"
-	"testing"
-
-	"github.com/spf13/afero"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/kubiyabot/cli/internal/config"
-	"github.com/kubiyabot/cli/internal/mcp"
-)
-
-// --- Removed Mocking for Kubiya Client ---
-
-func TestMcpInstallCommandExecute(t *testing.T) {
-	fs := afero.NewMemMapFs()
-	// Configure API key for auto-apply testing
-	// NOTE: Actual API call for teammates will NOT be mocked in this test setup.
-	// The test verifies install + initiation of auto-apply.
-	cfg := &config.Config{APIKey: "test-api-key"}
-	kubiyaDir, _ := mcp.GetKubiyaDir()
-	gatewayDir := filepath.Join(kubiyaDir, "mcp-gateway")
-	mcpConfigDir := filepath.Join(kubiyaDir, mcp.DefaultMcpDirName)
-
-	// --- Test Setup ---
-	startMockCommander()
-	defer stopMockCommander()
-
-	// --- Removed client mocking setup ---
-
-	// Mock filesystem for provider configs
-	require.NoError(t, fs.MkdirAll(mcpConfigDir, 0750))
-	// Add a compatible provider config (e.g., Cursor)
-	cursorConfigContent := `
-name: Cursor Test
-target_file: ~/.cursor/test_mcp.json
-template: |
-  {
-    "mcpServers": {
-      "kubiya-local": {
-        "command": "{{ .UvPath }}",
-        "args": ["run", "main.py"],
-        "env": {"KUBIYA_API_KEY": "{{ .ApiKey }}", "TEAMMATE_UUIDS": "{{ .TeammateUUIDsJSON }}"}
-      }
-    }
-  }
-`
-	require.NoError(t, afero.WriteFile(fs, filepath.Join(mcpConfigDir, "cursor_test.yaml"), []byte(cursorConfigContent), 0644))
-
-	// Add an OS-incompatible provider config (if testing on darwin, make it linux)
-	if runtime.GOOS == "darwin" {
-		incompatibleConfigContent := `
-name: Linux Only Test
-os: linux 
-target_file: ~/linux_only.json
-template: Error!
-`
-		require.NoError(t, afero.WriteFile(fs, filepath.Join(mcpConfigDir, "linux_only.yaml"), []byte(incompatibleConfigContent), 0644))
-	}
-
-	// Mock commands
-	mockGitSHA := "abcdef1234567890"
-	addMockCommand("git clone https://github.com/kubiyabot/mcp-gateway.git "+gatewayDir, "Cloning...\nSuccess!", nil)
-	addMockCommand("uv sync", "Synchronized environment.", nil)
-	addMockCommand("git rev-parse HEAD", mockGitSHA, nil)
-	// Mock LookPath for uv (needed by applyProviderConfiguration)
-	originalLookPath := execLookPath
-	execLookPath = func(file string) (string, error) {
-		if file == "uv" {
-			return "/path/to/mock/uv", nil
-		}
-		return originalLookPath(file) // Fallback to original for other commands like git
-	}
-	defer func() { execLookPath = originalLookPath }() // Restore original LookPath
-
-	// --- Execute Command ---
-	cmd := newMcpInstallCommand(cfg, fs)
-	out := new(bytes.Buffer)
-	errOut := new(bytes.Buffer)
-	cmd.SetOut(out)
-	cmd.SetErr(errOut)
-
-	err := cmd.Execute()
-	// We might expect an error here now if the real API call fails without mocking,
-	// OR the install command might gracefully handle teammate fetch errors.
-	// For now, let's assume the command proceeds even if teammate fetch fails,
-	// as the warning messages were added.
-	// require.NoError(t, err, "mcp install should succeed") // Comment out for now
-	output := out.String()
-
-	// --- Assertions ---
-	// Check basic install output messages
-	assert.Contains(t, output, "Starting MCP server installation")
-	assert.Contains(t, output, "Repository cloned successfully")
-	assert.Contains(t, output, "Dependencies installed successfully")
-	assert.Contains(t, output, "Storing current version information...")
-
-	// Check auto-apply initiation messages
-	assert.Contains(t, output, "Automatically applying default MCP configurations")
-	// Check teammate fetching attempt (might show error or success depending on env)
-	assert.Contains(t, output, "Fetching all teammates...")
-	assert.Contains(t, output, fmt.Sprintf("Checking providers in %s", mcpConfigDir))
-	assert.Contains(t, output, "Attempting to apply configuration for: cursor_test")
-	// Check that the configuration writing is attempted/succeeded for the compatible provider
-	assert.Contains(t, output, "Successfully applied MCP configuration for Cursor Test!")
-	if runtime.GOOS == "darwin" {
-		assert.Contains(t, output, "Attempting to apply configuration for: linux_only")
-		assert.Contains(t, output, "Skipping: Provider 'Linux Only Test' requires OS 'linux'")
-	}
-
-	// Check final success message
-	assert.Contains(t, output, "MCP server installation complete!")
-	assert.Contains(t, output, "Default configurations (if applicable) have been automatically applied.")
-
-	// Check recorded command calls (git clone, uv sync, git rev-parse - apply uses LookPath mock)
-	calls := getRecordedCommandCalls()
-	require.Len(t, calls, 3, "Expected 3 commands to be called (clone, sync, rev-parse)")
-	assert.Equal(t, "git", calls[0].Name)
-	assert.Equal(t, []string{"clone", mcpGatewayRepo, gatewayDir}, calls[0].Args)
-	assert.Equal(t, "uv", calls[1].Name)
-	assert.Equal(t, []string{"sync"}, calls[1].Args)
-	assert.Equal(t, gatewayDir, calls[1].Dir)
-	assert.Equal(t, "git", calls[2].Name)
-	assert.Equal(t, []string{"rev-parse", "HEAD"}, calls[2].Args)
-	assert.Equal(t, gatewayDir, calls[2].Dir)
-
-	// Check filesystem state (SHA file)
-	sha, err := mcp.ReadGatewaySHA(fs)
-	require.NoError(t, err, "Should be able to read gateway SHA after install")
-	assert.Equal(t, mockGitSHA, sha, "Stored SHA should match mock git output")
-
-	// Check filesystem state (Applied config file - Cursor)
-	homeDir, _ := os.UserHomeDir()
-	cursorTargetFile := filepath.Join(homeDir, ".cursor", "test_mcp.json") // Path from YAML
-	cursorConfigWritten, err := afero.Exists(fs, cursorTargetFile)
-	require.NoError(t, err)
-	assert.True(t, cursorConfigWritten, "Cursor config file should have been written at %s", cursorTargetFile)
-
-	// Verify content of applied config
-	contentBytes, err := afero.ReadFile(fs, cursorTargetFile)
-	require.NoError(t, err)
-	var appliedData map[string]interface{}
-	err = json.Unmarshal(contentBytes, &appliedData)
-	require.NoError(t, err, "Applied config should be valid JSON")
-
-	mcpServers, ok := appliedData["mcpServers"].(map[string]interface{}) // Adjusted based on mergeMcpJson logic
-	require.True(t, ok, "Applied config should contain 'mcpServers' map")
-	kubiyaLocal, ok := mcpServers["kubiya-local"].(map[string]interface{})
-	require.True(t, ok, "Applied config should contain 'kubiya-local' server")
-	env, ok := kubiyaLocal["env"].(map[string]interface{})
-	require.True(t, ok, "kubiya-local should contain 'env' map")
-	assert.Equal(t, cfg.APIKey, env["KUBIYA_API_KEY"].(string))
-	// Check teammate UUIDs JSON
-	expectedUUIDJSON, _ := json.Marshal([]string{"uuid-1", "uuid-2"})
-	assert.JSONEq(t, string(expectedUUIDJSON), env["TEAMMATE_UUIDS"].(string))
-	assert.Equal(t, "/path/to/mock/uv", kubiyaLocal["command"].(string))
-
-	// TODO: Test case where API key is not set - should skip apply
-	// TODO: Test case where ListTeammates fails - should warn and proceed
-	// TODO: Test case where no provider configs exist - should mention no providers applied
-
-	// Test install skip if directory exists (should not run apply step)
-	out.Reset()
-	errOut.Reset()
-	startMockCommander() // Reset mocks
-
-	err = cmd.Execute()
-	require.NoError(t, err, "mcp install should succeed even if dir exists")
-	output = out.String()
-	assert.Contains(t, output, "Target directory already exists. Installation skipped.")
-	assert.NotContains(t, output, "Automatically applying default MCP configurations")
-	assert.Empty(t, getRecordedCommandCalls(), "No commands should be called if dir exists")
-	stopMockCommander()
-
-}
-
-// TODO: Add tests for error cases:
-// - git not found
-// - uv not found
-// - git clone fails
-// - uv sync fails
-// - git rev-parse fails
-// - StoreGatewaySHA fails
-
-// Mocking helpers need to be adjusted/created if not already present
-// e.g., for execLookPath and kubiya.NewMockClient
